Fix calling `res.end` in `getInitialProps`
A common way to do redirections in `getInitialProps` is to use `res.writeHead` and then call `res.end`. When doing this the lambda handler would end up resolving twice and causing an error. This is because next checks if `res.finished` is set before ending the response, but `next-aws-cloudfront` never sets it. This adds `res.finished` and sets it to true in `res.end`.