README: clarify how to use `build.env`
I'm fairly new to next.js and was only aware of this approach of loading environment variables:
https://nextjs.org/docs/basic-features/environment-variables

I could not make it work with this servlerss-next.js as it seems that the only way to forward env vars is via the `next.config.js`. It took me way too long to find out. (Especially as `.env.local`/`.env.production` did work as expected). So I've decided to add this FAQ section.