Update README on lambda runtime and ISR
ISR is no longer alpha-only. Also, default runtime has changed and there's a fourth lambda function.

I do wonder about the image lambda: the readme says "Note that the sharp image library is built for NodeJS 12.x" – is that still true and does it mean the default for that should be reverted?